refactor(multimodal): split grpc multimodal.rs into a module directory
`multimodal.rs` had grown to ~2.65k lines -- the largest flat file under
routers/grpc (harmony/ and regular/ are already directories). Split it by
pipeline phase into a `multimodal/` module directory. This is a pure move
with no behavior change.

- detect:    modality detection + content extraction (chat + messages)
- config:    model config-file registry + per-router component bundle
- process:   fetch -> preprocess -> expand tokens -> build intermediate
- assemble:  intermediate -> backend-specific MultimodalData + per-item split
- serialize: tensor byte/dtype serialization (f32 -> bf16/f16)
- transport: SHM-vs-inline resolution + /dev/shm namespace verification

Tests move next to the code they cover. The crate-facing API is unchanged:
the same symbols are re-exported from mod.rs, and internal helpers are now
enforced-private (pub(super)) instead of implicitly module-scoped.
